+++ BelpostBot/src/main/java/by/nikiter/model/db/dao/UserDao.java
@@ -1,10 +1,8 @@
 package by.nikiter.model.db.dao;
 
-import by.nikiter.model.db.entity.TrackingEntity;
 import by.nikiter.model.db.entity.UserEntity;
 import org.hibernate.Session;
 
-import java.util.Iterator;
 import java.util.List;
 
 public class UserDao implements BasicDao<UserEntity,String> {
@@ -33,44 +31,36 @@ public class UserDao implements BasicDao<UserEntity,String> {
     }
 
     @Override
-    public void save(UserEntity user) {
+    public void saveOrUpdate(UserEntity user) {
         session.beginTransaction();
-        session.save(user);
-        session.getTransaction().commit();
-    }
-
-    @Override
-    public void update(UserEntity user) {
-        session.beginTransaction();
-        session.update(user);
+        session.saveOrUpdate(user);
         session.getTransaction().commit();
     }
 
     @Override
     public void delete(UserEntity user) {
         session.beginTransaction();
-        for (TrackingEntity tracking : user.getTrackings()) {
-            user.removeTracking(tracking);
-        }
+        user.removeAllTrackings();
         session.delete(user);
         session.getTransaction().commit();
     }
 
     @Override
     public void deleteById(String username) {
-        session.beginTransaction();
         UserEntity user = session.find(UserEntity.class, username);
-
-        user.removeAllTrackings();
-
-        session.delete(user);
-        session.getTransaction().commit();
+        if (user != null) {
+            session.beginTransaction();
+            user.removeAllTrackings();
+            session.delete(user);
+            session.getTransaction().commit();
+        }
     }
 
     @Override
-    public void merge(UserEntity user) {
+    public UserEntity merge(UserEntity user) {
         session.beginTransaction();
-        session.merge(user);
+        UserEntity userEntity = (UserEntity) session.merge(user);
         session.getTransaction().commit();
+        return userEntity;
     }
 }
